#1044 Rename Hartshorn Discovery to Hartshorn SPI
