+++ hartshorn-core/src/main/java/org/dockbox/hartshorn/application/environment/ContextualApplicationEnvironment.java
@@ -41,8 +41,8 @@ import org.dockbox.hartshorn.component.ComponentLocator;
 import org.dockbox.hartshorn.component.populate.ComponentInjectionPointsResolver;
 import org.dockbox.hartshorn.component.populate.MethodsAndFieldsInjectionPointResolver;
 import org.dockbox.hartshorn.context.ModifiableContextCarrier;
-import org.dockbox.hartshorn.discovery.DiscoveryService;
-import org.dockbox.hartshorn.discovery.ServiceDiscoveryException;
+import org.dockbox.hartshorn.spi.DiscoveryService;
+import org.dockbox.hartshorn.spi.ServiceDiscoveryException;
 import org.dockbox.hartshorn.inject.ComponentKeyResolver;
 import org.dockbox.hartshorn.inject.StandardAnnotationComponentKeyResolver;
 import org.dockbox.hartshorn.proxy.ProxyOrchestrator;
