+++ hartshorn-spi/src/main/java/org/dockbox/hartshorn/spi/NoAvailableImplementationException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2023 the original author or authors.
+ * Copyright 2019-2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.dockbox.hartshorn.discovery;
+package org.dockbox.hartshorn.spi;
 
 /**
  * Thrown when no implementation is available for a service. This may indicate the absence of a
