+++ hartshorn-spi/src/test/java/org/dockbox/hartshorn/spi/HelloWorldService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2023 the original author or authors.
+ * Copyright 2019-2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.dockbox.hartshorn.discovery;
+package org.dockbox.hartshorn.spi;
 
 @FunctionalInterface
 public interface HelloWorldService {
