+++ pom.xml
@@ -110,7 +110,7 @@
                 <module>hartshorn-bom</module>
                 <module>hartshorn-assembly/parent/pom.parent.xml</module>
                 <module>hartshorn-util</module>
-                <module>hartshorn-discovery</module>
+                <module>hartshorn-spi</module>
                 <module>hartshorn-proxy</module>
                 <module>hartshorn-proxy-javassist</module>
                 <module>hartshorn-introspect</module>
